requirements.txt
# requirements.txt - النسخة المعدلة (أبريل 2026)
# تم تثبيت الـ versions عشان نضمن استقرار التشغيل

# ═══ الـ Core ═══
telethon==1.36.0              # MTProto للـ userbot
python-telegram-bot==21.10    # Bot API (يدعم Bot API 9.6 / حقل style)
aiohttp==3.10.11              # HTTP async لتحميل الصور

# ═══ مساعدات ═══
deep-translator==1.11.4       # الترجمة (لو محتاجها)
pillow==11.1.0                # معالجة الصور
python-dotenv==1.0.1          # قراءة .env

# ═══ اختياري لتسريع Telethon ═══
# cryptg==0.5.0.post0         # uncomment لو محتاج تسريع التشفير (يحتاج compiler)
